requirements.txt
boto3==1.14.6
botocore==1.17.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
MouseInfo==0.1.3
numpy==1.23.2
pandas==1.4.4
psycopg2-binary==2.9.3
pdfminer==20191125
Pillow==9.2.0
psycopg2==2.9.3
PyAutoGUI==0.9.50
pycryptodome==3.15.0
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
wxPython==4.2.0
